Implemented the move to door capability.
Need to look at how the building drawing is assigned when creating the
drawing as the rows may not be displaying the Y axis but instead may be
displaying the X axis.

Also need to implement the show person method so that the person is
shown in the way they want in.